requirements.txt
python-dotenv==1.0.1
aiogram==3.6.0
motor==3.4.0
python-dateutil
pytest
pytest_asyncio
